+++ src/components/App/App.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import './App.css';
 
+import Articles from '../../assets/articles/articles.json';
+
 // Components
 import Header from '../Header';
 import Skills from '../Skills';
@@ -8,22 +10,61 @@ import Projects from '../Projects';
 import Work from '../Work';
 import Footer from '../Footer';
 import ErrorPage from '../ErrorPage';
+import Article from '../Article';
+
+// Import React Router
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route
+} from "react-router-dom"
+
+// Various Minor Components
+const Deafult = () => {
+  return (
+    <div className="App">
+      <Header />
+      <Skills />
+      <Projects />
+      <Work />
+      <Footer />
+    </div>
+  )
+}
 
 export default class App extends React.Component {
-  render() {
-    // Handle ErrorPages
-    const error = new URL(window.location.href).searchParams.get("error"); // Get the error URL parameter
-    const errors = ["400", "401", "403", "404", "500", "501", "502"];
-    if (errors.includes(error)) return <ErrorPage error={error} />
+  getArticles() {
+    let lst = [];
 
+    for (let [idx, i] of Articles.entries()) {
+      lst.push(
+        <Route key={idx} exact path={`/${i.link}`} component={
+          () => <div className="App"><Article title={i.title} body={i.body} /></div>
+        } />
+      )
+    }
+    return lst;
+  }
+
+  render() {
     return (
-      <div className="App">
-        <Header />
-        <Skills />
-        <Projects />
-        <Work />
-        <Footer />
-      </div>
+      <Router>
+        <Switch>
+          <Route exact path="/" component={Deafult} />
+
+          {this.getArticles()}
+
+          <Route exact path="/error/400" component={() => <ErrorPage error="400" />} />
+          <Route exact path="/error/401" component={() => <ErrorPage error="401" />} />
+          <Route exact path="/error/403" component={() => <ErrorPage error="403" />} />
+          <Route exact path="/error/404" component={() => <ErrorPage error="404" />} />
+          <Route exact path="/error/500" component={() => <ErrorPage error="500" />} />
+          <Route exact path="/error/501" component={() => <ErrorPage error="501" />} />
+          <Route exact path="/error/502" component={() => <ErrorPage error="502" />} />
+
+          <Route exact component={() => <ErrorPage error="404" />} />
+        </Switch>   
+      </Router>
     );
   }
 }